Write basic codes
