+++ src/Components/Edit/Edit.js
@@ -1,6 +1,7 @@
 import styled from "styled-components"
 import Header from "./Header"
 import Main from "./Main"
+import HighlightMenu from "./HighlightMenu"
 
 const EditStyle = styled.div`
   padding-top: 65px;
@@ -24,6 +25,7 @@ const Edit = ({
         keyDownEventListener={keyDownEventListener}
         checkMultiLineSelected={checkMultiLineSelected}
       />
+      <HighlightMenu />
     </EditStyle>
   )
 }
