+++ src/Components/Edit/HighlightMenu/Arrow.js
@@ -1,0 +1,5 @@
+import styled from "styled-components"
+
+const Arrow = styled.div``
+
+export default Arrow
