+++ src/Components/Edit/HighlightMenu/ButtonWrapper.js
@@ -1,0 +1,9 @@
+import styled from "styled-components"
+
+const ButtonWrapperStyle = styled.div``
+
+const ButtonWrapper = () => {
+  return <ButtonWrapperStyle>this is button wrapper</ButtonWrapperStyle>
+}
+
+export default ButtonWrapper
